# kv: pipeline replicated lock acquisition