Fixes #117978.

Builds upon the foundation laid in [#119975](https://github.com/cockroachdb/cockroach/pull/119975), [#119933](https://github.com/cockroachdb/cockroach/pull/119933), [#121065](https://github.com/cockroachdb/cockroach/pull/121065), and [#121086](https://github.com/cockroachdb/cockroach/pull/121086).

This commit completes the client-side handling of replicated lock acquisition pipelining. Replicated lock acquisition through `Get`, `Scan`, and `ReverseScan` requests now qualifies to be pipelined. The `txnPipeliner` is updated to track the strength associated with each in-flight write and pass that along to the corresponding `QueryIntentRequest`.

See [benchmark with TPC-C results here](https://github.com/cockroachdb/cockroach/pull/121088#issuecomment-2027848825).

Release note: None