@nvanbenschoten 2024-03-30
### Summary

For a multi-region cluster running TPC-C under Read Committed, this change provides a 75% speedup for the `delivery` transaction and a 38% speedup the `newOrder` transaction. 

For a single-region cluster running TPC-C under Read Committed, this change provides a 32% speedup for the `delivery` transaction, a 21% speedup for the `newOrder` transaction, and a 13% speedup for the `payment` transaction.

-----

### Impact on multi-region TPC-C

This change will be most impactful for clusters that with cross-region synchronous replication, where each round of raft takes O(10ms). We can see a significant improvement in multi-region clusters on Read Committed workloads that use SELECT FOR UPDATE and/or perform foreign key checks.

The following performance summary comes from a modified version of the `tpcc/headroom/isolation-level=read-committed/n4cpu16` roachtest which replicates across three regions, uses region survivability, and pins all leases to the region with the workload. Of note are the p50 latencies for the `delivery`, `newOrder`, and `payment` transactions.

**Before**
```
_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
 7200.0s        0         131480           18.3    381.5    385.9    402.7    419.4   1208.0  delivery

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
 7200.0s        0        1315851          182.8    172.6    176.2    184.5    226.5    738.2  newOrder

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
 7200.0s        0         131514           18.3      4.1      4.1      5.2      6.6    226.5  orderStatus

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
 7200.0s        0        1315369          182.7     71.7     71.3     79.7    142.6    671.1  payment

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
 7200.0s        0         131600           18.3      9.9      8.9     13.1     41.9    218.1  stockLevel
```

**After**
```
_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
 7200.0s        0         131705           18.3     92.1     96.5    100.7    113.2    369.1  delivery

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
 7200.0s        0        1318307          183.1    109.4    109.1    113.2    159.4    486.5  newOrder

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
 7200.0s        0         131764           18.3      4.5      4.7      6.0      7.1    184.5  orderStatus

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
 7200.0s        0        1318058          183.1     72.0     71.3     75.5    134.2    469.8  payment

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
 7200.0s        0         131696           18.3     10.1      9.4     13.1     37.7    234.9  stockLevel
```

`delivery` performs 10 sequential single-row SELECT FOR UPDATE statements (`selectNewOrder`), so pipelining this replicated lock acquisition is a big deal. With this change, its p50 latency from **385.9ms** to **96.5ms**, a **75%** improvement.

`newOrder` performs 1 multi-row SELECT FOR UPDATE statement (`selectStock`). It also performs 4 foreign key checks (1x in `insertOrder`, 1x in `insertNewOrder`, 2x in `insertOrderLine`). With this change, its p50 latency from **176.2ms** to **109.1ms**, a **38%** improvement.

`payment` performs performs 2 foreign key checks (2x in `insertHistory`). We don't see a meaningful improvement for this test case, but do see one below.


### Impact on single-region TPC-C

This change will be most impactful for clusters that with cross-region replication. However, we do see an improvement in single-region clusters on Read Committed workloads that use SELECT FOR UPDATE and/or perform foreign key checks. This is because pipelining allows the disk writes for replicated locks to be parallelized.

The following performance summary comes from the `tpcc/headroom/isolation-level=read-committed/n4cpu16` roachtest. Of note are the p50 latencies for the `delivery`, `newOrder`, and `payment` transactions.

**Before**
```
_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
 7200.0s        0         132250           18.4     57.9     58.7     71.3     83.9    436.2  delivery

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
 7200.0s        0        1322298          183.7     25.5     25.2     31.5     41.9    318.8  newOrder

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
 7200.0s        0         132210           18.4      5.4      5.5      7.1      8.4    192.9  orderStatus

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
 7200.0s        0        1321391          183.5     12.5     12.6     15.7     22.0    251.7  payment

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
 7200.0s        0         132120           18.3     10.9     11.0     15.2     19.9     75.5  stockLevel
```

**After**
```
_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
 7200.0s        0         132228           18.4     38.9     39.8     48.2     54.5    251.7  delivery

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
 7200.0s        0        1322691          183.7     20.0     19.9     24.1     29.4    218.1  newOrder

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
 7200.0s        0         132230           18.4      4.8      5.0      6.3      7.3    104.9  orderStatus

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
 7200.0s        0        1321782          183.6     11.0     11.0     13.6     16.8    209.7  payment

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
 7200.0s        0         132061           18.3     10.0     10.0     14.2     17.8     54.5  stockLevel
```

With this change, the `delivery` transaction's p50 latency drops from **58.7ms** to **39.8ms**, a **32%** speedup.

With this change, the `newOrder` transaction's p50 latency drops from **25.2ms** to **19.9ms**, a **21%** speedup.

With this change, the `payment` transaction's p50 latency drops from **12.6ms** to **11.0ms**, a **13%** speedup.